test(policy): centralize doctor fixtures
