+++ extensions/policy/src/doctor/register.base.test-utils.ts
@@ -20,32 +20,18 @@ import {
   ctx,
   repairCtx,
   runPolicyChecks,
+  runPolicyChecksFixture,
   runDeniedChannelRepair,
   runPolicyRepairCheck,
-  describe0BeforeEach0,
-  describe0AfterEach1,
+  setupPolicyDoctorTest,
+  teardownPolicyDoctorTest,
+  writePolicyFixture,
 } from "./register.test-harness.js";
 
-async function writePolicyFixture(...json: Parameters<typeof JSON.stringify>): Promise<string> {
-  const [policy] = json;
-  const configPath = join(workspaceDir, "openclaw.jsonc");
-  await fs.writeFile(configPath, "{}", "utf-8");
-  await fs.writeFile(
-    join(workspaceDir, "policy.jsonc"),
-    typeof policy === "string" ? policy : JSON.stringify(...json),
-    "utf-8",
-  );
-  return configPath;
-}
-
-async function runPolicyChecksFixture(policy: unknown) {
-  return runPolicyChecks(ctx(await writePolicyFixture(policy), cfgWithPolicy()));
-}
-
 describe("registerPolicyDoctorChecks", () => {
-  beforeEach(describe0BeforeEach0);
+  beforeEach(setupPolicyDoctorTest);
 
-  afterEach(describe0AfterEach1);
+  afterEach(teardownPolicyDoctorTest);
 
   it("allows scoped overrides that are stricter than top-level policy", async () => {
     const result = await runPolicyChecksFixture({
